+++ HISTORY.md
@@ -41,6 +41,8 @@ _????-??-??_
  * Fix infinite recursion in `Octree` when the number of identical points
    exceeds `maxLeafSize` (#4020).
 
+ * Add YOLOv3Tiny for object detection (#4023).
+
  * Fix thread-specific random seed initialization (#4027).
 
 ## mlpack 4.6.2